add raw post #+URL: Marcin Borkowski: 2017-03-04 Bookmarking directories in Eshell [[http://mbork.pl/Homepage][Homepage]] [[http://mbork.pl/Recent_Changes][Recent Changes]]
Search:

[[http://mbork.pl?search=%222017-03-04+Bookmarking+directories+in+Eshell%22][2017-03-04 Bookmarking directories in Eshell]]

A friend of mine showed me a nifty trick, basically [[http://jeroenjanssens.com/2013/08/16/quickly-navigate-your-filesystem-from-the-command-line.html][bookmarking directories in bash]]. While this is nice, I often use Eshell, and I wanted a similar thing there. (Notice that while Emacs bookmarks do work in Dired, they won’t let you bookmark a directory in Eshell!) Bonus points for sharing the bookmark database (which is basically a directory called ~/.marks/ containing symlinks to bookmarked directories).

Well, here it is. Notice the (a bit misleadingly named) function file-symlink-p, which returns the target of the symlink if the file with the given name is a symlink and nil otherwise. Also, the directory name is hardcoded, and I didn’t bother to add any error detection – but this would be easy.

(defun eshell/jump (mark)
  "Jump to a directory symlinked to by a file called ~/.marks/MARK."
  (eshell/cd (file-symlink-p (concat "~/.marks/" mark))))

This is nice already, but lacks one important thing: autocompletion. It turns out that this is not difficult at all:

(defun pcomplete/jump ()
  "Complete a command that wants a name of a file in ~/.marks."
  (pcomplete-here* (directory-files "~/.marks/")))

And that’s pretty much it!

On the other hand, it would be probably more Emacs-y to make Emacs bookmarks support Eshell. Still, this solution has an advantage of working both in Eshell and in bash. Moreover, one could add support also for the mark function, which wouldn’t be too difficult.

[[http://mbork.pl/CategoryEnglish][CategoryEnglish]], [[http://mbork.pl/CategoryBlog][CategoryBlog]], [[http://mbork.pl/CategoryEmacs][CategoryEmacs]]

[[http://mbork.pl/Comments_on_2017-03-04_Bookmarking_directories_in_Eshell][Comments on 2017-03-04 Bookmarking directories in Eshell]] [[http://mbork.pl?action=password][This page is read-only]] [[http://mbork.pl?action=history;id=2017-03-04_Bookmarking_directories_in_Eshell][View other revisions]] [[http://mbork.pl?action=admin;id=2017-03-04_Bookmarking_directories_in_Eshell][Administration]]Last edited 2017-03-04 17:06 UTC by [[http://mbork.pl/Marcin_Borkowski][Marcin Borkowski]]